opfor_uniforms - move csat fatigues config into separate files
