+++ addons/opfor_uniforms/CfgWeapons.hpp
@@ -3,198 +3,6 @@ class CfgWeapons
 	class ItemCore;
 	class UniformItem;
 	class Uniform_Base;
-	class tmtm_u_csatFatigues_atacsAu: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (ATACS AU)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_atacsAu_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_universal_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_atacsAu";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_atacsFg: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (ATACS FG)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_atacsFg_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_atacsFg";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_black: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (Black)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_black_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_black";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_blue: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (Blue)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_blue_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_blue";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_emr: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (EMR)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_emr_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_emr";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_grey: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (Grey)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_grey_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_grey";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_greyBlack: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (Grey/Black)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_greyBlack_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_greyBlack";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_ldf: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (LDF)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_ldf_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_ldf";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_mtp: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (MTP)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_mtp_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_mtp";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_multicam: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (Multicam)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_multicam_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_multicam";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_natoWdl: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (NATO Woodland)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_natoWdl_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_natoWdl";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_surpat: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (SURPAT)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_surpat_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_surpat";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
-	
-	class tmtm_u_csatFatigues_yellow: Uniform_Base
-	{
-		scope=2;
-		displayName="CSAT Fatigues (Yellow)";
-		picture="\tmtm\pack\opfor_uniforms\Data\UI\tmtm_icon_combatFatiguesCsat_yellow_ca.paa";
-		model="\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
-		class ItemInfo: UniformItem
-		{
-			uniformModel="-";
-			uniformClass="tmtm_csatFatigues_yellow";
-			containerClass="Supply40";
-			mass=80;
-		};
-	};
+
+	#include "include/csatFatigues/CfgWeapons.hpp"
 };
